Move lookup_sample() from AnophelesBase to AnophelesSampleMetadata
